Fix missing SideOnly annotations
We've assumed that Forge adds its annotation everywhere via its patches but it
actually only adds them if a given file needs to have a patch applied for
reasons other than just the SideOnly annotation.
This would leave some classes/fields/methods without annotation which resulted
in an exception when deserializing horses on 1.12.2 (because it references a
class which in turn has methods referencing the dedicated server class which is
server-only).

Instead, we now convert all Fabric annotations (added during merging) to the
Forge ones and remove duplicates (where patches also added them).